+++ packages/find-and-replace/lib/project/results-model.js
@@ -1,6 +1,7 @@
 const _ = require('underscore-plus')
 const {Emitter, TextEditor, Range} = require('atom')
 const escapeHelper = require('../escape-helper')
+const path = require('path');
 
 class Result {
   static create(result) {
@@ -342,16 +343,26 @@ module.exports = class ResultsModel {
     this.emitter.emit('did-set-result', {filePath, result})
   }
 
+  // Ensure the given file path is suitable for inclusion in the current
+  // results view by checking it against any path patterns.
+  //
+  // Ordinary project-wide search should already produce results that match our
+  // paths pattern; but we also search all modified buffers in the workspace in
+  // manual fashion, so we need this check to to be able to filter those
+  // buffers properly.
   shouldAddResult(filePath) {
-    // Ensure the given file path is suitable for inclusion in the current
-    // results view by checking it against any path patterns.
-    if (!this.findOptions.pathsPattern) return true;
-    const searchPaths = this.pathsArrayFromPathsPattern(this.findOptions.pathsPattern);
-    return atom.workspace.filePathMatchesPatterns(filePath, searchPaths);
+    // return true;
+    let { pathsPattern } = this.findOptions
+    if (!pathsPattern) return true
+
+    const searchPaths = this.pathsArrayFromPathsPattern(pathsPattern)
+    return atom.workspace.filePathMatchesPatterns(filePath, searchPaths)
   }
 
   addResult(filePath, result) {
-    if (!this.shouldAddResult(filePath, result)) return;
+    if (!this.shouldAddResult(filePath, result)) {
+      return
+    }
 
     this.pathCount++
     this.matchCount += result.matches.length