+++ packages/settings-view/spec/package-card-spec.js
@@ -3,6 +3,7 @@ const path = require('path');
 const PackageCard = require('../lib/package-card');
 const PackageManager = require('../lib/package-manager');
 const SettingsView = require('../lib/settings-view');
+const {shell} = require('electron');
 
 describe("PackageCard", function() {
   const setPackageStatusSpies = function(opts) {
@@ -75,6 +76,28 @@ describe("PackageCard", function() {
     expect(card.refs.updateButton).toBeVisible();
     expect(card.refs.updateButton.textContent).toContain('Update to 1.2.0');
   });
+
+  it("shows a badge", function () {
+    const pack = {
+      badges: [{
+        link: 'https://example.com',
+        title: 'Archived',
+        text: 'Source code has been archived',
+        type: 'warn'
+      }],
+      name: 'something',
+      version: '1.0.0',
+      latestVersion: '1.0.0'
+    };
+    card = new PackageCard(pack, new SettingsView(), packageManager);
+
+    spyOn(shell, 'openExternal');
+    jasmine.attachToDOM(card.element);
+    let badge = card.element.querySelector('.badge');
+    expect(badge).toExist();
+    badge?.click();
+    expect(shell.openExternal).toHaveBeenCalled();
+  })
 
   it("shows the author details", function() {
     const authorName = "authorName";